tests/wpt/tests/fedcm/fedcm-store.https.html

<!DOCTYPE html>
<title>Federated Credential Management API store() tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {fedcm_test,
        request_options_with_mediation_required,
        fedcm_get_and_select_first_account} from "./support/fedcm-helper.sub.js";

fedcm_test(async t => {
  const cred = await fedcm_get_and_select_first_account(t, request_options_with_mediation_required());
  return promise_rejects_dom(t, "NotSupportedError", navigator.credentials.store(cred));
}, "navigator.credentials.store() with an identity credential returns NotSupportedError");

</script>
